pyproject.toml
[project]
name = "healthcare-ach-monorepo"
version = "0.1.0"
description = "Multi-service healthcare ACH/EFT processing monorepo"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111.0",
  "pydantic>=2.7.0",
  "sqlalchemy>=2.0.0",
  "azure-servicebus>=7.12.0",
  "uvicorn[standard]>=0.30.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
